Refactored MantaObject to an interface and renamed original MantaObject to MantaMetadata.
